requirements.txt
# Enhanced Agentic Workflow Architecture Dependencies

# Core AI Framework
pydantic-ai==0.0.13
langgraph==0.2.44
pydantic==2.9.2

# LLM Providers
openai==1.57.2
anthropic==0.40.0
google-generativeai==0.8.3

# Database and Vector Search
supabase==2.8.1
redis==5.2.0
pgvector==0.3.6

# Web Framework and API
fastapi==0.115.6
uvicorn==0.32.1
streamlit==1.41.0
httpx==0.28.1

# Monitoring and Logging
logfire==0.54.0
structlog==24.4.0

# Development Tools
ruff==0.8.4
mypy==1.13.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# CLI and UI
rich==13.9.4
typer==0.13.1
click==8.1.7

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
asyncio-mqtt==0.16.2
websockets==13.1

# Docker and Deployment
docker==7.1.0
docker-compose==1.29.2

# MCP Protocol
mcp==1.0.0

# Additional FastAPI dependencies
websockets==13.1
sse-starlette==2.1.3
jinja2==3.1.4
python-multipart==0.0.19

# Data visualization
plotly==5.22.0
pandas==2.2.3
